third_party/WebKit/LayoutTests/http/tests/media/autoplay/document-user-activation-iframe-delegation-no-gesture.html

<!DOCTYPE html>
<html>
<title>Test that autoplay is still disabled if there is no gesture, but delegation is switched on.</title>
<iframe allowedactivationdelegation="media" src="http://localhost:8000/media/autoplay/resources/nested-iframe-1a.html"></iframe>
<script>
  if (window.testRunner) {
    testRunner.clearBackForwardList();
    testRunner.waitUntilDone();
  }
</script>
</html>
